+++ src/persistence/index.ts
@@ -1,3 +1,4 @@
 export * from './projectPersistence'
 export * from './projectArchive'
+export * from './projectWorkspace'
 export * from './projectSchemas'